requirements.txt
# requirements.txt
torch==2.0.1
torch-geometric==2.3.1
networkx==3.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
flask==2.3.2
web3==6.5.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.65.0
joblib==1.3.1